+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ✅ Added
 
-- Added `StreamMessageInput.sendMessageKey` and `StreamMessageInput.clearQuotedMessageKey` to customize the keys used to send and clear the quoted message.
+- Added `StreamMessageInput.sendMessageKeyPredicate` and `StreamMessageInput.clearQuotedMessageKeyPredicate` to customize the keys used to send and clear the quoted message.
 
 ## 5.1.0